requirements.txt
youtube_dl==2021.12.17
discord.py==2.1.0
PyNaCl==1.5.0
ffmpeg==1.4
requests==2.28.1
psutil==5.9.4
validators==0.20.0
